fix print10 to support long int

+++ trunk/metal-cell/include/uprint.h
@@ -4,7 +4,7 @@
 void lcd_putc(char data);
 
 char hex2c(char hex);
-void print10(unsigned short n);
+void print10(unsigned long n);
 void printLL(unsigned long n, char dot, char prec);
